fix(exception-capture): key rate limiter on root cause of chained exceptions
PostHog groups exceptions by $exception_list[0].type, which is the root
cause (exceptions_from_error_tuple reverses the walked chain), not the
wrapping exception. Walk the chain so e.g. `RuntimeError from
ZeroDivisionError` is keyed on ZeroDivisionError, matching server-side
grouping and posthog-js.

Also move the ExceptionCapture integration tests out of
test_bucketed_rate_limiter.py into test_exception_capture.py so each
file covers a single unit.
