+++ posthog/test/test_bucketed_rate_limiter.py
@@ -215,98 +215,3 @@ def test_thread_safety_allows_exactly_bucket_size_minus_one():
         t.join()
 
     assert len(allowed) == 49
-
-
-def test_exception_capture_rate_limiting_is_disabled_by_default():
-    from posthog.exception_capture import ExceptionCapture
-
-    client = MagicMock()
-    capture = ExceptionCapture(client)
-    try:
-        assert capture._rate_limiter is None
-
-        for _ in range(100):
-            capture.capture_exception((ValueError, ValueError("boom"), None))
-        assert client.capture_exception.call_count == 100
-    finally:
-        capture.close()
-
-
-def test_exception_capture_default_configuration_when_enabled():
-    from posthog.exception_capture import ExceptionCapture
-
-    capture = ExceptionCapture(MagicMock(), rate_limiting_enabled=True)
-    try:
-        assert capture._rate_limiter._bucket_size == 50
-        assert capture._rate_limiter._refill_rate == 10
-        assert capture._rate_limiter._refill_interval == 10
-    finally:
-        capture.close()
-
-
-def test_exception_capture_rate_limiting_is_configurable():
-    from posthog.exception_capture import ExceptionCapture
-
-    capture = ExceptionCapture(
-        MagicMock(),
-        rate_limiting_enabled=True,
-        bucket_size=3,
-        refill_rate=2,
-        refill_interval_seconds=5,
-    )
-    try:
-        assert capture._rate_limiter._bucket_size == 3
-        assert capture._rate_limiter._refill_rate == 2
-        assert capture._rate_limiter._refill_interval == 5
-    finally:
-        capture.close()
-
-
-def test_client_passes_rate_limiter_configuration_through():
-    from posthog.client import Client
-
-    client = Client(
-        "phc_test",
-        sync_mode=True,
-        disabled=True,
-        enable_exception_autocapture=True,
-        enable_exception_autocapture_rate_limiting=True,
-        exception_autocapture_bucket_size=3,
-        exception_autocapture_refill_rate=2,
-        exception_autocapture_refill_interval_seconds=5,
-    )
-    try:
-        limiter = client.exception_capture._rate_limiter
-        assert limiter._bucket_size == 3
-        assert limiter._refill_rate == 2
-        assert limiter._refill_interval == 5
-    finally:
-        client.shutdown()
-
-
-def test_exception_capture_rate_limits_per_exception_type():
-    from posthog.exception_capture import ExceptionCapture
-
-    client = MagicMock()
-    capture = ExceptionCapture(client, rate_limiting_enabled=True, bucket_size=10)
-    try:
-
-        def exc_info(error):
-            try:
-                raise error
-            except type(error):
-                import sys
-
-                return sys.exc_info()
-
-        for _ in range(15):
-            capture.capture_exception(exc_info(ValueError("boom")))
-
-        # bucket size 10 -> 9 captured, the rest rate limited
-        assert client.capture_exception.call_count == 9
-
-        # a different exception type has its own bucket
-        capture.capture_exception(exc_info(ZeroDivisionError("zero")))
-        assert client.capture_exception.call_count == 10
-    finally:
-        capture.close()
